+++ packages/widgetbook/lib/src/knobs/knobs.dart
@@ -6,5 +6,6 @@ export 'color_knob.dart';
 export 'double_input_knob.dart';
 export 'double_slider_knob.dart';
 export 'knob.dart';
+export 'knobs_registry.dart';
 export 'list_knob.dart';
 export 'string_knob.dart';